requirements.txt
pydantic~=2.9.2
fastapi~=0.115.5
mediapipe~=0.10.18
numpy~=1.26.4
matplotlib~=3.9.2
mysql-connector-python~=9.1.0
pydantic-settings~=2.6.1
bcrypt~=4.2.1
starlette~=0.41.2
pillow~=11.0.0
jwt~=1.3.1
passlib~=1.7.4
bcrypt==4.0.1